+++ internal/services/pinger/internal/handler/command_ping.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 
 	"github.com/number571/go-peer/pkg/logger"
@@ -10,7 +11,7 @@ import (
 	"github.com/number571/hidden-lake/internal/utils/chars"
 	http_logger "github.com/number571/hidden-lake/internal/utils/logger/http"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
-	hlk_request "github.com/number571/hidden-lake/pkg/network/request"
+	"github.com/number571/hidden-lake/pkg/api/services/pinger/request"
 )
 
 func HandleCommandPingAPI(
@@ -27,12 +28,7 @@ func HandleCommandPingAPI(
 			return
 		}
 
-		req := hlk_request.NewRequestBuilder().
-			WithMethod(http.MethodGet).
-			WithHost(hls_settings.CAppShortName).
-			WithPath(hls_settings.CPingPath).
-			Build()
-
+		req := request.NewPingRequest()
 		resp, err := pHlkClient.FetchRequest(pCtx, pR.URL.Query().Get("friend"), req)
 		if err != nil {
 			pLogger.PushErro(logBuilder.WithMessage("fetch_request"))
@@ -40,16 +36,16 @@ func HandleCommandPingAPI(
 			return
 		}
 
-		if resp.GetCode() != http.StatusOK {
+		if code := resp.GetCode(); code != http.StatusOK {
 			pLogger.PushErro(logBuilder.WithMessage("status_error"))
-			_ = api.Response(pW, http.StatusInternalServerError, "failed: status error")
+			_ = api.Response(pW, http.StatusTeapot, fmt.Sprintf("failed: status %d", code))
 			return
 		}
 
 		respBody := string(resp.GetBody())
 		if chars.HasNotGraphicCharacters(respBody) {
 			pLogger.PushErro(logBuilder.WithMessage("invalid_response"))
-			_ = api.Response(pW, http.StatusInternalServerError, "failed: invalid response")
+			_ = api.Response(pW, http.StatusServiceUnavailable, "failed: invalid response")
 			return
 		}